backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.232
supabase==2.18.1
openai==1.107.1
python-dotenv==1.0.0
reportlab==4.0.7
jinja2==3.1.2
python-jose==3.3.0
bcrypt==4.1.2
aiofiles==23.2.1
scikit-image==0.22.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.3
requests==2.31.0
